requirements.txt
apache-beam[gcp]==2.52.0
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==3.0.1
cloudpickle==2.2.1
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
fastavro==1.7.0
fasteners==0.18
google-api-core==2.11.0
google-apitools==0.5.31
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.1
grpcio-status==1.48.2
hdfs==2.7.0
httplib2==0.20.4
idna==3.4
mpmath==1.3.0
numpy==1.22.4
oauth2client==4.1.3
objsize==0.6.1
orjson==3.8.5
overrides==6.5.0
packaging==23.0
proto-plus==1.22.2
protobuf==3.20.3
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pymongo==3.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.14
zstandard==0.19.0
torch==2.0.1
